+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Semantic versioning
 Version latest (0.5.0?, 2019-09-24?)
 
 - ...
+- enh: factor out code into `ApiMan` (for listening on SQS and calling AWS API Gateway endpoint) and new class EmailMan
+    - this makes the code much simpler with `listen_sqs` call instead of all the internals showing up in `tagsSuggestAdvanced`
 
 
 Version 0.11.0 (2019-10-15)