+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Version latest (0.5.0?, 2019-09-24?)
     - this makes the code much simpler with `listen_sqs` call instead of all the internals showing up in `tagsSuggestAdvanced`
 - feat: http requests to aws api gateway to be sigv4-signed with aws keys (in order to do clean authentication on http)
     - this alleviates the need to do further authentication via sqs for example (where sqs is limited to the source account by iam policy)
+    - WIP
 
 
 Version 0.11.0 (2019-10-15)